Updated predictive
Made a whole bunch of changes to the predictive. It should maybe be able
to make predictions now, but that hasn't been debugged mostly because
there's no good way to show the aerial robot right now.

Oh and it compiles again. You can run it and watch the pretty robots
dance.